Move `getFirstMaterialTypeFromFilters` to 'search-result/helper.ts'
